LayoutTests/http/tests/site-isolation/opened-window-length.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifies that windows opened with same-origin and cross-origin iframes have a correct length.");
jsTestIsAsync = true;

var lengths = [];
onmessage = (event) => {
    if (event.data == "#w1")
        lengths.push(w1.length);
    else if (event.data == "#w2")
        lengths.push(w2.length);

    if (lengths.length == 2) {
        shouldBe("lengths", "[2, 2]");
        finishJSTest();
    }
}

function runTest() {
    w1 = window.open("http://localhost:8000/site-isolation/resources/post-message-to-opener-on-iframes-load.html#w1");
    w2 = window.open("http://127.0.0.1:8000/site-isolation/resources/post-message-to-opener-on-iframes-load.html#w2");
}
</script>
<body onload="runTest()">
</body>
