LayoutTests/http/tests/site-isolation/opened-window-length-expected.txt

Verifies that windows opened with same-origin and cross-origin iframes have a correct length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lengths is [2, 2]
PASS successfullyParsed is true

TEST COMPLETE
